[refactor] #22 국회의원Util 코드 일부 국회의원Service 클래스로 이동
